# packages/api-contracts/openapi/workplace.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts/{postId}": {
      "patch": {
        "description": "Updates post content or visibility for the author.",
        "operationId": "PostsController_updatePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated post payload"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update post",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "description": "Deletes a post owned by the current user.",
        "operationId": "PostsController_deletePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete result status"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete post",
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostsController_getPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Canonical post payload"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get canonical post detail",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/comments/posts/{postId}": {
      "post": {
        "operationId": "CommentsController_createComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "get": {
        "operationId": "CommentsController_listComments",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/comments/{commentId}": {
      "patch": {
        "operationId": "CommentsController_updateComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      },
      "delete": {
        "operationId": "CommentsController_deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
    "/comments/{commentId}/report": {
      "post": {
        "operationId": "CommentsController_reportComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
...
    "/comments/reports/{reportId}/resolve": {
      "post": {
        "operationId": "CommentsController_resolveReport",
        "parameters": [
          {
            "name": "reportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveCommentReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Workplace API",
    "description": "Workplace backend API documentation",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 2000
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "followers",
              "private"
            ]
          }
        }
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCommentReportDto": {
        "type": "object",
        "properties": {}
      },
      "ResolveCommentReportDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}